// codegen/crm/extensions/accounting/model/accountingExtensionCustomer.ts
/**
 * Accounting Extension
 * These APIs allow you to interact with HubSpot\'s Accounting Extension. It allows you to: * Specify the URLs that HubSpot will use when making webhook requests to your external accounting system. * Respond to webhook calls made to your external accounting system by HubSpot 
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { Address } from './address';

/**
* Representation of a customer in the external accounting system.
*/
export class AccountingExtensionCustomer {
    /**
    * The customer\'s email address
    */
    'emailAddress'?: string;
    /**
    * The customer\'s full name
    */
    'name': string;
    /**
    * The ID of the customer in the external accounting system.
    */
    'id': string;
    'billingAddress'?: Address;
    /**
    * The ISO 4217 currency code that represents the currency the customer should be billed in.
    */
    'currencyCode'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "emailAddress",
            "baseName": "emailAddress",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "billingAddress",
            "baseName": "billingAddress",
            "type": "Address"
        },
        {
            "name": "currencyCode",
            "baseName": "currencyCode",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return AccountingExtensionCustomer.attributeTypeMap;
    }
}