// codegen/crm/extensions/accounting/model/createUserAccountRequestExternal.ts
/**
 * Accounting Extension
 * These APIs allow you to interact with HubSpot\'s Accounting Extension. It allows you to: * Specify the URLs that HubSpot will use when making webhook requests to your external accounting system. * Respond to webhook calls made to your external accounting system by HubSpot 
 *
 * The version of the OpenAPI document: v3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

/**
* Information about the account in your external account system.
*/
export class CreateUserAccountRequestExternal {
    /**
    * The id of the account in your system.
    */
    'accountId': string;
    /**
    * The name of the account in your system. This is normally the name visible to your users.
    */
    'accountName': string;
    /**
    * The default currency that this account uses.
    */
    'currencyCode': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accountId",
            "baseName": "accountId",
            "type": "string"
        },
        {
            "name": "accountName",
            "baseName": "accountName",
            "type": "string"
        },
        {
            "name": "currencyCode",
            "baseName": "currencyCode",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateUserAccountRequestExternal.attributeTypeMap;
    }
}